tilt: replace outputter with PipelineState
